requirements.txt
streamlit>=1.32.0
openai>=1.14.0
python-dotenv>=1.0.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-docx>=1.1.0
fpdf2>=2.7.6
pandas>=2.0.0
